+++ src/ts/Components/Panes.ts
@@ -1,4 +1,4 @@
-import { Game } from "./Game";
+import { Game } from "../Game";
 
 export class Panes {
   Game: Game;
