+++ src/ts/Components/ProjectileManager.ts
@@ -1,10 +1,10 @@
-import { Game } from "./Game";
+import { Game } from "../Game";
 
-import { AudioLoader } from "./Libraries/Loaders";
-import { Direction } from "./Libraries/Direction";
-import { Collisions } from "./Libraries/Collisions";
+import { AudioLoader } from "../Libraries/Loaders";
+import { Direction } from "../Libraries/Direction";
+import { Collisions } from "../Libraries/Collisions";
 
-import { Projectile, ProjectileState } from "./Projectiles";
+import { Projectile, ProjectileState } from "../Projectiles/Projectiles";
 
 export class ProjectileManager {
   Game: Game;
