+++ src/ts/Components/Viewport.ts
@@ -1,12 +1,12 @@
-import { Game, GameState } from "./Game";
+import { Game, GameState } from "../Game";
 
-import { Direction } from "./Libraries/Direction";
-import { Collisions } from "./Libraries/Collisions";
+import { Direction } from "../Libraries/Direction";
+import { Collisions } from "../Libraries/Collisions";
 
-import { World } from "./Map/World";
-import { Scene } from "./Map/Scene";
-import { Passage } from "./Map/Passage";
-import { Cell } from "./Map/Cell";
+import { World } from "../Map/World";
+import { Scene } from "../Map/Scene";
+import { Passage } from "../Map/Passage";
+import { Cell } from "../Map/Cell";
 import { EnemyManager } from "./EnemyManager";
 
 export class Viewport {
