+++ src/ts/Game.ts
@@ -1,17 +1,17 @@
 import { StateObserver } from "./Libraries/Observers";
 
-import { BrickCollection } from "./Bricks";
+import { BrickCollection } from "./Bricks/Bricks";
 import { Map } from "./Map/Map";
-import { Viewport } from "./Viewport";
-import { Inventory } from "./Inventory";
-import { Player } from "./Player";
-import { Sword } from "./Sword";
-import { EnemyManager } from "./EnemyManager";
-import { ProjectileManager } from "./ProjectileManager";
-import { ItemManager } from "./ItemManager";
-import { EventManager } from "./EventManager";
-import { Hud } from "./Hud";
-import { Panes } from "./Panes";
+import { Viewport } from "./Components/Viewport";
+import { Inventory } from "./Components/Inventory";
+import { Player } from "./Components/Player";
+import { Sword } from "./Components/Sword";
+import { EnemyManager } from "./Components/EnemyManager";
+import { ProjectileManager } from "./Components/ProjectileManager";
+import { ItemManager } from "./Components/ItemManager";
+import { EventManager } from "./Components/EventManager";
+import { Hud } from "./Components/Hud";
+import { Panes } from "./Components/Panes";
 import { SplashScreen } from "./Screens/SplashScreen";
 import { GameOverScreen } from "./Screens/GameOverScreen";
 import { WinScreen } from "./Screens/WinScreen";
